replace NDEBUG by macro defined in auto-generated version header
the version header will in addition serve as build log, info on compiler-tools is saved there, too.
this obviously introduces some changes in the makefile, as the header has to be generated (in the
bjuild/ directory), the install target changes and the like.

besides, the uninstall target in the makefile is fixed, too.