implement automatic simplification of trigonometric power constructs in products
implemented by replacing sine and cosine expressions by temporary variables, and tangent expressions
by the quotient of two temporary variables. this is then simplified as usual, and the result is
transformed back to expressions with functions. as this procedure differs from cohen's algorithm, it
has to implemented independently and thus before the usual automatic simplification procedure.

this commit includes some refactoring of the ProductSimpl class.