handle symmetries of (inverse) trigonometric functions
this implies:

sin(-a) = -sin(a)
cos(-a) = cos(a)
tan(-a) = -tan(a)

asin(-a) = asin(a)
acos(-a) = pi - acos(a)
atan(-a) = -atan(a)

atan2(-b, a) = -atan(b/a) if b/a numerically evaluable and -b/a < 0

the minus sign is recognized for products, e.g. -2*a*(b + c), for sums if all summands have negative
factors, and for numerically evaluable arguments. that should cover all possibilities so far.