move string representation from BasePtr into Base class
This speeds up running tests from debug builds, but the main reason for
this is, that shifting debug strings around should not be handled in the
reference-counting wrapper class. Now, Base is equipped with a protected
method that can be called from its derived classes once they're
initialized (that means, in the ctor body). Setting up the the debug
string earlier (e.g. in Base ctor) is impossible, because printing
expressions obviously needs the full state of Base-children classes.