use simple, custom logging to reduce library dependencies
A new Logger interface is added, which, by default, prints warning or
more severe messages to stdout. The header is exposed to the public API
such that applications can provide their own logging class. As an
example, logging output of unit tests is managed by a simple new Logger
subclass.

Using this facade-like logging mechanism keeps things as simple as
possible and eliminates the need for an external library. The formatting
of log messages is still important and must be addressed: here, a
formatting tool that mixes std::printf and std::ostream operators has
been added and is invoked via the logging macros used throughout the
source.