always shift atan2 result into [0, 2*pi] range
Note that this mechanism is actually quite a horrible artifact - we are
changing co-domain and semantics of a well-established function
available in any programming environment. Not sure how I came up with
this in the first place, but the whole NumTrigoSimpl class is so
atrocious and needs a serious rewrite, that it's impossible to tackle it
for now - hence the quickfix.